QCamera2: Fix manual ISO control
Our proprietary camera libraries were not built by the OEM with manual
ISO in mind, so there are many strange requirements that need to be in
place in order for a manual ISO setting to stick. These weird
requirements are all located in the backend, so this patch involves
sending multiple commands to the backend in order to attain the
special condition where manual ISO settings are honored.

Signed-off-by: freshgiammi <rengucci.gianmarco@gmail.com>